moved the example to the example folder

+++ examples/neural_network/sae_mnist.py
@@ -8,7 +8,7 @@ mnist = fetch_mldata('MNIST original')
 X, y = mnist.data, mnist.target
 random.seed(100)
 indices = np.array(random.sample(range(70000), 1000))
-X, y = X[indices].astype('float32'), y[indices]
+X, y = X[indices].astype('float64'), y[indices]
 # For SAE, feature values in the range [0, 1] is necessary
 X = X / 255
 sae = SAE(
@@ -20,6 +20,10 @@ sae = SAE(
 sae_features = sae.fit_transform(X)
 clf = SGDClassifier(random_state=3)
 clf.fit(X, y)
+#Should get a score of 0.943
 print 'SGD on raw pixels score: ', clf.score(X, y)
 clf.fit(sae_features, y)
+#Should get a score of 0.982
 print 'SGD on extracted features score: ', clf.score(sae_features, y)
+
+
